STUB of adding custom fields to uploaded files
(this commit mixes other branches, sorry)
Right now the patch allows to:
* add a custom footer to the image (view file)
* adds a "title" input field

The field is uploaded using `uploadExtraData`, handled by FileController and stored into session, using the file name as index (this
is not functionally because the user could upload two files with the same file name).
After that, when `module->attachFiles` is called, the data is taken from the session and stored to the File model in the additional_info
field.
When the user loads the model the data is loaded using `getInitialPreviewThumbTags()`.

Several problems persists:
  the update. Additional logic should be implemented in order to take care of the form data and update the model
  the uploaded file (you will have index 0 for the existing file and index 1 for the new one, but the `uploadedFile` instance has just 1
  element)

Requires the following in the _form, but most of this code should be taken to the module.
    <?=
    \nemmo\attachments\components\AttachmentsInput::widget([
        'id' => 'file-input',
        'model' => $model,
        'options' => [
            'multiple' => true,
        ],
        'pluginOptions' => [
            'maxFileCount' => 10,
            'initialPreview' => $model->isNewRecord ? [] : $model->getInitialPreview(),
            'initialPreviewConfig' => $model->isNewRecord ? [] : $model->getInitialPreviewConfig(),
            'initialPreviewThumbTags' => $model->isNewRecord ? [] : $model->getInitialPreviewThumbTags(),
            'layoutTemplates' => ['footer' => $this->render("@vendor/nemmo/yii2-attachments/views/previewTemplates/_image")],
            'previewThumbTags' => ['{TAG_TITLE}' => ''],
            'uploadExtraData' => new yii\web\JsExpression('function() {  // callback example
        var out = {}, key, i = 0;
        $(".kv-input:visible").each(function() {
            $el = $(this);
            out["title"+i] = $el.val();
            i++;
        });
        return out;}'),
        ]
    ]);
    ?>